refactor tests and add github action
