--- .travis.yml
@@ -1,20 +1,0 @@
-language: go
-go:
-  # older versions are not tested, due to SQL driver compatibility
-  - 1.12.x
-  - 1.13.x
-  - 1.14.x
-  - 1.15.x
-
-services:
-  - mysql
-  - postgresql
-
-install:
-  - go get github.com/go-sql-driver/mysql
-  - go get github.com/lib/pq
-
-script:
-  - test -z "$(go fmt ./...)" # fail if not formatted properly
-  - go vet
-  - make test # tests postgres and mysql databases
