+++ go.mod
@@ -1,3 +1,8 @@
 module github.com/DATA-DOG/go-txdb
 
 go 1.12
+
+require (
+	github.com/go-sql-driver/mysql v1.7.1
+	github.com/lib/pq v1.10.9
+)
